+++ common/src/main/java/dev/nolij/zume/common/util/IFileWatcher.java
@@ -1,0 +1,9 @@
+package dev.nolij.zume.common.util;
+
+public interface IFileWatcher {
+	
+	void lock() throws InterruptedException;
+	boolean tryLock();
+	void unlock();
+	
+}