+++ legacy/src/main/java/dev/nolij/zume/legacy/LegacyZume.java
@@ -19,7 +19,7 @@ public class LegacyZume implements ClientModInitializer, IZumeImplementation {
 		
 		Zume.LOGGER.info("Loading Legacy Zume...");
 		
-		Zume.init(this, FabricLoader.getInstance().getConfigDir().resolve(Zume.CONFIG_FILE_NAME).toFile());
+		Zume.init(this, FabricLoader.getInstance().getConfigDir());
 	}
 	
 	@Override