+++ lexforge16/src/main/java/dev/nolij/zume/lexforge16/LexZume16.java
@@ -14,8 +14,6 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.loading.FMLEnvironment;
 import net.minecraftforge.fml.loading.FMLPaths;
 
-import java.io.File;
-
 @Mod(Zume.MOD_ID)
 public class LexZume16 implements IZumeImplementation {
 	
@@ -27,7 +25,7 @@ public class LexZume16 implements IZumeImplementation {
 		
 		LexZume16ConfigScreen.register();
 		
-		Zume.init(this, new File(FMLPaths.CONFIGDIR.get().toFile(), Zume.CONFIG_FILE_NAME));
+		Zume.init(this, FMLPaths.CONFIGDIR.get());
 		if (Zume.disabled)
 			return;